create trash list and recyclerview on fragment home
