+++ app/src/main/AndroidManifest.xml
@@ -25,16 +25,16 @@
         tools:targetApi="31" >
 
         <activity
-            android:name=".ui.welcome.WelcomeActivity"
+            android:name=".presentation.activity.WelcomeActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.inputdata.InputDataActivity"
+            android:name=".presentation.activity.InputDataActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.onboard.OnboardPageActivity"
+            android:name=".presentation.activity.OnboardPageActivity"
             android:exported="false" />
         <activity
-            android:name=".splash.SplashScreenActivity"
+            android:name=".presentation.activity.SplashScreenActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -43,7 +43,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.activity.MainActivity"
             android:exported="false" />
     </application>
