+++ app/src/main/java/com/trashcare/user/presentation/activity/OnboardPageActivity.kt
@@ -1,14 +1,11 @@
 @file:Suppress("DEPRECATION")
 
-package com.trashcare.user.ui.onboard
+package com.trashcare.user.presentation.activity
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.Window
-import android.view.WindowManager
 import com.trashcare.user.databinding.ActivityOnboardPageBinding
-import com.trashcare.user.ui.inputdata.InputDataActivity
 
 class OnboardPageActivity : AppCompatActivity() {
