+++ app/src/main/java/com/trashcare/user/presentation/fragment/CameraFragment.kt
@@ -1,23 +1,21 @@
-package com.trashcare.user.ui.fragment
+package com.trashcare.user.presentation.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.addCallback
 import com.trashcare.user.R
 
 
-@Suppress("DEPRECATION")
-class HomeFragment : Fragment() {
+class CameraFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
-
+        return inflater.inflate(R.layout.fragment_camera, container, false)
     }
+
 }
