+++ app/src/main/java/com/trashcare/user/presentation/fragment/HomeFragment.kt
@@ -1,0 +1,37 @@
+package com.trashcare.user.presentation.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import com.trashcare.user.R
+import com.trashcare.user.databinding.FragmentHomeBinding
+
+
+@Suppress("DEPRECATION")
+class HomeFragment : Fragment() {
+
+    private var _binding: FragmentHomeBinding?= null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        // Inflate the layout for this fragment
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
+
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.btnSendTrash.setOnClickListener {
+            findNavController().navigate(R.id.home_to_camera)
+        }
+    }
+}
