--- app/src/main/java/com/trashcare/user/ui/fragment/CameraFragment.kt
@@ -1,21 +1,0 @@
-package com.trashcare.user.ui.fragment
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.trashcare.user.R
-
-
-class CameraFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_camera, container, false)
-    }
-
-}
