+++ app/src/main/res/drawable/rectangle_17.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="328dp"
+    android:height="40dp"
+    android:viewportWidth="328"
+    android:viewportHeight="40"
+    >
+    <group>
+        <clip-path
+            android:pathData="M9 0H319C323.971 0 328 4.02944 328 9V31C328 35.9706 323.971 40 319 40H9C4.02944 40 0 35.9706 0 31V9C0 4.02944 4.02944 0 9 0Z"
+            />
+        <path
+            android:pathData="M0 0V40H328V0"
+            android:fillColor="#2EB5B5B5"
+            />
+    </group>
+</vector>
