+++ app/src/main/res/layout/activity_onboard_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.onboard.OnboardPageActivity">
+    tools:context=".presentation.activity.OnboardPageActivity">
 
     <ImageView
         android:id="@+id/imageView"
