+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/green_500"
-    tools:context=".splash.SplashScreenActivity">
+    tools:context=".presentation.activity.SplashScreenActivity">
 
     <TextView
         android:id="@+id/trashcare"
