+++ app/src/main/res/layout/activity_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.welcome.WelcomeActivity">
+    tools:context=".presentation.activity.WelcomeActivity">
 
     <View
         android:id="@+id/rectangle_3"
