+++ app/src/main/res/layout/fragment_camera.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.CameraFragment">
+    tools:context=".presentation.fragment.CameraFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
