+++ app/src/main/res/layout/fragment_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.HistoryFragment">
+    tools:context=".presentation.fragment.HistoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
