+++ app/src/main/res/layout/item_trash.xml
@@ -1,0 +1,94 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_margin="20dp"
+    app:cardCornerRadius="10dp">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <ImageView
+            android:id="@+id/iv_trash"
+            android:layout_width="92dp"
+            android:layout_height="110dp"
+            android:src="@drawable/plastic_bottles_img"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+            android:id="@+id/tv_trash"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/poppins_bold"
+            android:layout_marginStart="25dp"
+            android:text="@string/title_trash_botle_plasstic"
+            android:textSize="13sp"
+            app:layout_constraintStart_toEndOf="@+id/iv_trash"
+            tools:ignore="MissingConstraints" />
+
+        <TextView
+            android:id="@+id/tv_trash_desc"
+            android:layout_width="179dp"
+            android:layout_height="32dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginStart="25dp"
+            android:fontFamily="@font/poppins_regular"
+            android:textSize="12sp"
+            android:text="@string/txt_desc_trash"
+            app:layout_constraintStart_toEndOf="@+id/iv_trash"
+            app:layout_constraintTop_toBottomOf="@+id/tv_trash" />
+
+
+        <Button
+            android:id="@+id/btn_detail_trash"
+            android:layout_width="72dp"
+            android:layout_height="30dp"
+            android:layout_marginTop="8dp"
+            android:fontFamily="@font/poppins_regular"
+            android:text="@string/txt_detail_button_trash"
+            android:textSize="8sp"
+            app:layout_constraintEnd_toEndOf="@+id/tv_trash_desc"
+            app:layout_constraintTop_toBottomOf="@+id/tv_trash_desc"
+            tools:ignore="SmallSp,TouchTargetSizeCheck" />
+
+
+        <ImageButton
+            android:id="@+id/btn_plus"
+            android:layout_width="20dp"
+            android:layout_height="20dp"
+            android:layout_marginBottom="10dp"
+            android:src="@drawable/baseline_add_box_24"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="@+id/tv_trash_desc"
+            tools:ignore="ContentDescription,TouchTargetSizeCheck,SpeakableTextPresentCheck" />
+
+        <TextView
+            android:id="@+id/tv_result"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="2dp"
+            android:layout_marginBottom="10dp"
+            tools:text="0"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/btn_plus" />
+
+        <ImageButton
+            android:id="@+id/btn_min"
+            android:layout_width="20dp"
+            android:layout_height="20dp"
+            android:layout_marginStart="2dp"
+            android:layout_marginBottom="10dp"
+            android:src="@drawable/baseline_min_check_box_24"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/tv_result"
+            tools:ignore="ContentDescription,TouchTargetSizeCheck,SpeakableTextPresentCheck" />
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.cardview.widget.CardView>
