+++ app/src/main/res/navigation/bottom_navigation.xml
@@ -7,19 +7,25 @@
 
     <fragment
         android:id="@+id/action_home"
-        android:name="com.trashcare.user.ui.fragment.HomeFragment"
+        android:name="com.trashcare.user.presentation.fragment.HomeFragment"
         android:label="Home"
-        tools:layout="@layout/fragment_home"/>
+        tools:layout="@layout/fragment_home">
+
+        <action
+            android:id="@+id/home_to_camera"
+            app:destination="@id/action_camera"/>
+
+    </fragment>
 
     <fragment
         android:id="@+id/action_camera"
-        android:name="com.trashcare.user.ui.fragment.CameraFragment"
+        android:name="com.trashcare.user.presentation.fragment.CameraFragment"
         android:label="Camera"
         tools:layout="@layout/fragment_camera"/>
 
     <fragment
         android:id="@+id/action_history"
-        android:name="com.trashcare.user.ui.fragment.HistoryFragment"
+        android:name="com.trashcare.user.presentation.fragment.HistoryFragment"
         android:label="History"
         tools:layout="@layout/fragment_history"/>
